Refactor i18n exports and translation analysis
